requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache
redis==5.0.1

# ML and NLP
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
spacy==3.7.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Vector Database
pymilvus==2.3.4

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.0

# Logging
loguru==0.7.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utils
python-dateutil==2.8.2

# Frontend (Streamlit)
streamlit==1.28.1
plotly==5.17.0
streamlit-option-menu==0.3.6
